fixed file tree (don't kow how)
fixed the file tree hierarchy not working like it should
the fix is really strange though, it only worked before due to
a typo in the old code. I removed the line that I had "fixed" and
now it works as it did before.

Added option to choose the name of the database file (defaults to :memory:)

added sorter.py file for debug purposes